more fixes

+++ xv6-riscv/kernel/exec.c
@@ -117,7 +117,7 @@ exec(char *path, char **argv)
   // }
 
   //writing bar for bar of old xv6 code
-  uint oldsz=0;
+  uint64 oldsz=0;
   for (int i=0; i< MAXTHREADS;i++){
     uint64 sz1;
     if((sz1 = uvmalloc(pagetable, sz, sz + 2*PGSIZE, PTE_W )) == 0)
